change background colors, remove unused components
