--- client/src/components/Buttons/SignupButton.tsx
@@ -1,25 +1,0 @@
-import { useAuth0 } from "@auth0/auth0-react";
-import { Button } from "@mantine/core";
-import React from "react";
-
-export const SignupButton: React.FC = () => {
-  const { loginWithRedirect } = useAuth0();
-
-  const handleSignUp = async () => {
-    await loginWithRedirect({
-      appState: {
-        returnTo: "/timer",
-      },
-      authorizationParams: {
-        prompt: "login",
-        screen_hint: "signup",
-      },
-    });
-  };
-
-  return (
-    <Button onClick={handleSignUp} variant="outline">
-      Sign Up
-    </Button>
-  );
-};
