+++ client/src/index.css
@@ -66,3 +66,7 @@ button:focus-visible {
     background-color: #f9f9f9;
   }
 } */
+
+body {
+  background-color: #4B2842;
+}
