+++ client/src/pages/Timer/ProjectButton.tsx
@@ -2,8 +2,8 @@ import { DEFAULT_THEME, Modal, Button, Menu, TextInput, ColorPicker, Text, Stack
 import { useDisclosure } from '@mantine/hooks';
 import { IconFolderOpen, IconPlus, IconPointFilled } from "@tabler/icons-react";
 import { useEffect, useState } from "react";
-import { Project } from "../../../classes/models";
-import { getProjects, postProject } from "../../../classes/HTTPhelpers";
+import { Project } from "../../classes/models";
+import { getProjects, postProject } from "../../classes/HTTPhelpers";
 
 interface NewProjectModalProps {
   opened: boolean;
